+++ raven_front/lib/pages/wallet/asset_details/asset_details_widget.dart
@@ -72,59 +72,3 @@ class _TransactionsState extends State<Transactions> {
         .ofMediaHeight(context);
   }
 }
-
-class MetadataView extends StatelessWidget {
-  const MetadataView({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    Asset securityAsset = assetDetailsBloc.security.asset!;
-
-    var chilren = <Widget>[];
-    if (securityAsset.primaryMetadata == null &&
-        securityAsset.hasData &&
-        securityAsset.data!.isIpfs) {
-      final height =
-          (assetDetailsBloc.scrollObserver.value.ofMediaHeight(context) + 32) /
-              2;
-      return Container(
-        alignment: Alignment.topCenter,
-        height: height,
-        child: Padding(
-          padding: EdgeInsets.only(top: 16),
-          child: components.buttons.actionButtonSoft(
-            context,
-            label: 'View Data',
-            onPressed: () => components.message.giveChoices(
-              context,
-              title: 'View Data',
-              content: 'View data in external browser?',
-              behaviors: {
-                'CANCEL': Navigator.of(context).pop,
-                'BROWSER': () {
-                  Navigator.of(context).pop();
-                  launch('https://ipfs.io/ipfs/${securityAsset.metadata}');
-                },
-              },
-            ),
-          ),
-        ),
-      );
-    } else if (securityAsset.primaryMetadata == null) {
-      chilren = [SelectableText(securityAsset.metadata)];
-    } else if (securityAsset.primaryMetadata!.kind == MetadataType.ImagePath) {
-      chilren = [
-        Image.file(AssetLogos()
-            .readImageFileNow(securityAsset.primaryMetadata!.data ?? ''))
-      ];
-    } else if (securityAsset.primaryMetadata!.kind == MetadataType.JsonString) {
-      chilren = [SelectableText(securityAsset.primaryMetadata!.data ?? '')];
-    } else if (securityAsset.primaryMetadata!.kind == MetadataType.Unknown) {
-      chilren = [
-        SelectableText(securityAsset.primaryMetadata!.metadata),
-        SelectableText(securityAsset.primaryMetadata!.data ?? '')
-      ];
-    }
-    return ListView(padding: EdgeInsets.all(10.0), children: chilren);
-  }
-}
